+++ src/lib/internal.h
@@ -9,6 +9,7 @@ extern "C" {
 #include "builddef.h"
 #include "compat/compat.h"
 #include "notcurses/notcurses.h"
+#include "notcurses/direct.h"
 
 // KEY_EVENT is defined by both ncurses.h and wincon.h. since we don't use
 // either definition, kill it before inclusion of ncurses.h.
@@ -271,6 +272,15 @@ typedef struct nctabbed {
   nctabbed_options opts; // copied in nctabbed_create()
 } nctabbed;
 
+// various moving parts within a notcurses context (and the user) might need to
+// access the stats object, so throw a lock on it. we don't want the lock in
+// the actual structure since (a) it's usually unnecessary and (b) it breaks
+// memset() and memcpy().
+typedef struct ncsharedstats {
+  pthread_mutex_t lock;
+  ncstats s;
+} ncsharedstats;
+
 typedef struct ncdirect {
   ncpalette palette;         // 256-indexed palette can be used instead of/with RGB
   FILE* ttyfp;               // FILE* for output tty
@@ -279,6 +289,7 @@ typedef struct ncdirect {
   uint16_t stylemask;        // current styles
   bool initialized_readline; // have we initialized Readline?
   uint64_t flags;            // copied in ncdirect_init() from param
+  ncsharedstats stats;       // stats! not as broadly used as in notcurses
 } ncdirect;
 
 // Extracellular state for a cell during the render process. There is one
@@ -329,15 +340,6 @@ typedef struct ncpile {
   sprixel* sprixelcache;      // list of sprixels
 } ncpile;
 
-// various moving parts within a notcurses context (and the user) might need to
-// access the stats object, so throw a lock on it. we don't want the lock in
-// the actual structure since (a) it's usually unnecessary and (b) it breaks
-// memset() and memcpy().
-typedef struct ncsharedstats {
-  pthread_mutex_t lock;
-  ncstats s;
-} ncsharedstats;
-
 // the standard pile can be reached through ->stdplane.
 typedef struct notcurses {
   ncplane* stdplane; // standard plane, covers screen
@@ -375,7 +377,6 @@ typedef struct notcurses {
   int loglevel;
   ncpalette palette; // 256-indexed palette can be used instead of/with RGB
   bool palette_damage[NCPALETTESIZE];
-  unsigned stdio_blocking_save; // was stdio blocking at entry? restore on stop.
   uint64_t flags;  // copied from notcurses_options
 } notcurses;
 
@@ -1743,6 +1744,44 @@ tty_check(int fd){
   return fd >= 0 ? 1 : 0;
 #endif
   return isatty(fd);
+}
+
+// attempt to cancel the specified thread (not an error if we can't; it might
+// have already exited), and then join it (an error here is propagated).
+static inline int
+cancel_and_join(const char* name, pthread_t tid, void** res){
+  if(pthread_cancel(tid)){
+    logerror("couldn't cancel %s thread\n", name); // tid might have died
+  }
+  if(pthread_join(tid, res)){
+    logerror("error joining %s thread\n", name);
+    return -1;
+  }
+  return 0;
+}
+
+static inline int
+emit_scrolls(const tinfo* ti, int count, fbuf* f){
+  if(count > 1){
+    const char* indn = get_escape(ti, ESCAPE_INDN);
+    if(indn){
+      if(fbuf_emit(f, tiparm(indn, count)) < 0){
+        return -1;
+      }
+      return 0;
+    }
+  }
+  const char* ind = get_escape(ti, ESCAPE_IND);
+  if(ind == NULL){
+    ind = "\v";
+  }
+  while(count > 0){
+    if(fbuf_emit(f, ind) < 0){
+      return -1;
+    }
+    --count;
+  }
+  return 0;
 }
 
 #undef ALLOC