+++ src/poc/bitmapstates.c
@@ -149,6 +149,7 @@ wipebitmap(struct notcurses* nc){
 int main(void){
   struct notcurses_options opts = {
     .loglevel = NCLOGLEVEL_TRACE,
+    .flags = NCOPTION_DRAIN_INPUT,
   };
   struct notcurses* nc = notcurses_core_init(&opts, NULL);
   if(notcurses_check_pixel_support(nc) < 1){