+++ src/poc/scroll.c
@@ -5,8 +5,10 @@
 int main(void){
   setlocale(LC_ALL, "");
   struct notcurses_options opts = {
-    .flags = NCOPTION_INHIBIT_SETLOCALE | NCOPTION_NO_ALTERNATE_SCREEN
-              | NCOPTION_PRESERVE_CURSOR,
+    .flags = NCOPTION_INHIBIT_SETLOCALE
+              | NCOPTION_NO_ALTERNATE_SCREEN
+              | NCOPTION_PRESERVE_CURSOR
+              | NCOPTION_DRAIN_INPUT,
   };
   struct notcurses* nc = notcurses_core_init(&opts, NULL);
   if(nc == NULL){